download data as csv, because ECDC can't output JSON properly ><
