+++ config.json
@@ -1,7 +1,7 @@
 {
     "port": 3000,
     "listeningInterface": "127.0.0.1",
-    "dataUri": "https://opendata.ecdc.europa.eu/covid19/casedistribution/json",
+    "dataBaseUri": "https://opendata.ecdc.europa.eu/covid19/casedistribution",
     "db": {
         "database": "db/covid19cases.db"
     },
