+++ package.json
@@ -7,6 +7,7 @@
   },
   "dependencies": {
     "cookie-parser": "~1.4.4",
+    "csv-parse": "^4.8.8",
     "debug": "~2.6.9",
     "ejs": "~2.6.1",
     "express": "~4.16.1",
